+++ compconfig/install.sh
@@ -1,3 +1,0 @@
-# bash commands for installing your package
-BUILD_NUM=1
-conda install -c pslmodels -c conda-forge nodejs taxbrain "paramtools>=0.7.0"