+++ cs-config/cs_config/functions.py
@@ -11,7 +11,8 @@ from .helpers import (convert_defaults, convert_adj, TCDIR,
 from .outputs import create_layout, aggregate_plot
 from taxbrain import TaxBrain, report
 from dask import delayed, compute
-from collections import defaultdict
+from collections import defaultdict, OrderedDict
+from marshmallow import fields
 
 
 AWS_ACCESS_KEY_ID = os.environ.get("AWS_ACCESS_KEY_ID", "")
@@ -39,6 +40,13 @@ class BehaviorParams(paramtools.Parameters):
     defaults = behavior_params
 
 
+def get_version():
+    model_versions_str = ""
+    for model, version in TaxBrain.VERSIONS.items():
+        model_versions_str += f"{model}: {version}\n"
+    return model_versions_str
+
+
 def get_inputs(meta_params_dict):
     """
     Return default parameters for Tax-Brain
@@ -47,30 +55,25 @@ def get_inputs(meta_params_dict):
     metaparams.adjust(meta_params_dict)
 
     policy_params = TCParams()
+    policy_params.set_state(
+        year=metaparams.year.tolist(),
+        data_source=metaparams.data_source,
+    )
+
+    filtered_pol_params = OrderedDict()
+    for k, v in policy_params.dump().items():
+        if k =="schema" or v.get("section_1", False):
+            filtered_pol_params[k] = v
+
     behavior_params = BehaviorParams()
 
     default_params = {
-        "policy": policy_params.specification(
-            meta_data=True,
-            include_empty=True,
-            year=metaparams.year,
-            data_source=metaparams.data_source,
-            use_full_sample=metaparams.use_full_sample,
-            serializable=True
-        ),
-        "behavior": behavior_params.specification(
-            meta_data=True,
-            include_empty=True,
-            serializable=True
-        )
+        "policy": filtered_pol_params,
+        "behavior": behavior_params.dump()
     }
-    meta = metaparams.specification(
-        meta_data=True,
-        include_empty=True,
-        serializable=True
-    )
+    meta = metaparams.dump()
 
-    return meta, default_params
+    return {"meta_parameters": meta, "model_parameters": default_params}
 
 
 def validate_inputs(meta_params_dict, adjustment, errors_warnings):
@@ -92,7 +95,7 @@ def validate_inputs(meta_params_dict, adjustment, errors_warnings):
     errors_warnings["behavior"]["errors"].update(behavior_params.errors)
 
     if policy_params.errors or behavior_params.errors:
-        return errors_warnings
+        return {"errors_warnings": errors_warnings}
 
     # try to parse to the correct Tax-Calculator format.
     try:
@@ -105,9 +108,9 @@ def validate_inputs(meta_params_dict, adjustment, errors_warnings):
                                   meta_params.year.tolist()),
             "behavior": convert_behavior_adj(adjustment["behavior"])
         }
-        res = errors_warnings, tc_adj
+        res = {"errors_warnings": errors_warnings, "custom_adjustment": tc_adj}
     except Exception:
-        res = errors_warnings
+        res = {"errors_warnings": errors_warnings}
         print("Error parsing:", adjustment)
         traceback.print_exc()
 
@@ -198,12 +201,9 @@ def run_model(meta_params_dict, adjustment):
             }
         )
     agg_output, table_output = create_layout(results, start_year, end_year)
-    model_versions_str = ""
-    for model, version in TaxBrain.VERSIONS.items():
-        model_versions_str += f"{model}: {version}\n"
+
     comp_outputs = {
         "renderable": [agg_plot, agg_output, table_output],
-        "model_version": model_versions_str,
         "downloadable": downloadable
     }
     return comp_outputs