+++ cs-config/install.sh
@@ -1,0 +1,2 @@
+# bash commands for installing your package
+conda install -c pslmodels -c conda-forge nodejs "taxbrain>=2.3.2" "taxcalc>=2.5.0" "paramtools>=0.10.2" pypandoc